1-last_digit.c

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -13,7 +13,7 @@ int main(void)
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
 	/* your code goes there */
-	int ld = n % 10;
+	float ld = n % 10;
 	if (ld > 5)
 	{
 		printf("Last digit of %d is %d and is greater than 5\n", n, ld);
